requirements.txt
dspy-ai>=2.4.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0

